swiggy_rag/requirements.txt
pypdf
tiktoken
chromadb
numpy
openai
python-dotenv